handle delete tag and category

+++ app/Http/Controllers/CategoriesController.php
@@ -99,6 +99,11 @@ class CategoriesController extends Controller
      */
     public function destroy(Category $category)
     {
+        if($category->posts->count() > 0) {
+            session()->flash('error', 'Category cannot be delete because it has some posts.');
+
+            return redirect()->back();
+        }
         $category->delete();
 
         session()->flash('success', 'Successfully Deleted');
